Change flip_length to ulong (from size_t) to avoid issues on Windows

+++ src/main.c
@@ -35,7 +35,7 @@ typedef struct {
 	triple t;
 	ulong freq;
 	ulong plot_freq;
-	size_t flip_length; 
+	ulong flip_length; 
 } sim_params;
